+++ src/test/java/com/jinju/jinjuwiki/domain/auth/service/AuthServiceTest.java
@@ -12,9 +12,6 @@ import com.jinju.jinjuwiki.domain.auth.dto.response.SignupResponse;
 import com.jinju.jinjuwiki.domain.auth.repository.EmailVerificationRepository;
 import com.jinju.jinjuwiki.global.error.BusinessException;
 import com.jinju.jinjuwiki.global.error.ErrorCode;
-import com.jinju.jinjuwiki.domain.user.entity.User;
-import com.jinju.jinjuwiki.domain.user.entity.UserRole;
-import com.jinju.jinjuwiki.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +19,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
@@ -31,9 +27,6 @@ class AuthServiceTest {
 
     @Autowired
     private AuthService authService;
-
-    @Autowired
-    private UserRepository userRepository;
 
     @Autowired
     private EmailVerificationRepository emailVerificationRepository;
@@ -99,46 +92,6 @@ class AuthServiceTest {
                 .isInstanceOf(BusinessException.class)
                 .extracting("errorCode")
                 .isEqualTo(ErrorCode.INVALID_LOGIN);
-    }
-
-    @Test
-    @DisplayName("DB unique 제약으로 이메일이 중복되어도 데이터 무결성 예외가 발생한다.")
-    void duplicateEmailCanFailAtDatabaseLevel() {
-        userRepository.save(User.builder()
-                .email("dbdup@test.com")
-                .password("encoded-password")
-                .nickname("dbUser1")
-                .role(UserRole.USER)
-                .build());
-
-        assertThatThrownBy(() -> userRepository.saveAndFlush(User.builder()
-                .email("dbdup@test.com")
-                .password("encoded-password")
-                .nickname("dbUser2")
-                .role(UserRole.USER)
-                .build()))
-                .isInstanceOf(DataIntegrityViolationException.class);
-    }
-
-    @Test
-    @DisplayName("이메일 인증이 완료되지 않으면 회원가입할 수 없다.")
-    void signupFailWhenEmailNotVerified() {
-        assertThatThrownBy(() -> authService.signup(new SignupRequest("unverified@test.com", "password123", "user3")))
-                .isInstanceOf(BusinessException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.EMAIL_NOT_VERIFIED);
-    }
-
-    @Test
-    @DisplayName("인증코드 검증에 성공하면 인증 완료 상태가 된다.")
-    void verifyCodeSuccess() {
-        authService.sendVerificationCode(new EmailVerificationSendRequest("verify@test.com"));
-        String code = emailVerificationRepository.findByEmail("verify@test.com").orElseThrow().getCode();
-
-        var response = authService.verifyCode(new EmailVerificationVerifyRequest("verify@test.com", code));
-
-        assertThat(response.verified()).isTrue();
-        assertThat(response.email()).isEqualTo("verify@test.com");
     }
 
     private void verifyEmail(String email) {
